Use Go binary to get epochtime
I have benchmarked:
- 'date +%s.%N' -> 1.3ms
- 'resh-get-epochtime' -> 2.1ms
- 'date --version' -> 1.1ms

This means that it's better to run the Go binary once than
running 'date' multiple times to check that it is installed
and then to get the time.

Also it's just much simpler and robust to have one binary
instead of shell function with multiple if-else branches.